Ramdisk: Profiles: Make battery settings my own
Both Joe and I found that the current battery settings were not that
great. So I redid them with my own. These settings are the closest
I can get to truly hampering performance without causing serious
stutter, which we do not want. No input boost, sluggish scaling on
both big and little, underclocking both cores, and offlining two big
cores should result in some serious reduction in power consumption.

I probably would not recommend running this as a daily profile but it
should get the job done when needed.

Thanks to szale9001 on XDA for the tip for offlining the big cores:
https://forum.xda-developers.com/showpost.php?p=71602280&postcount=6183

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>